+++ monitor/price/snapshot_test.go
@@ -69,7 +69,10 @@ func TestEmpty(t *testing.T) {
 
 	// Create a new market and restore into it
 	riskModel, settings := createPriceMonitorDeps(t, ctrl)
-	pm2, err := price.NewMonitorFromSnapshot(state1, settings, riskModel)
+
+	statevar := mocks.NewMockStateVarEngine(ctrl)
+	statevar.EXPECT().RegisterStateVariable(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).Times(1)
+	pm2, err := price.NewMonitorFromSnapshot("marketID", "assetID", state1, settings, riskModel, statevar, logging.NewTestLogger())
 	require.NoError(t, err)
 	require.NotNil(t, pm2)
 
@@ -113,7 +116,9 @@ func TestChangedState(t *testing.T) {
 	state := pm1.GetState()
 
 	riskModel, settings := createPriceMonitorDeps(t, ctrl)
-	pm2, err := price.NewMonitorFromSnapshot(state, settings, riskModel)
+	statevar := mocks.NewMockStateVarEngine(ctrl)
+	statevar.EXPECT().RegisterStateVariable(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).Times(1)
+	pm2, err := price.NewMonitorFromSnapshot("marketID", "assetID", state, settings, riskModel, statevar, logging.NewTestLogger())
 	require.NoError(t, err)
 	require.NotNil(t, pm2)